day10: fix 7J and FL counting giving different results
The issue was that we didn't handle start tile properly. It was marked as S and we completely ignored it. In reality it can be any of the pipes and we may need to count it.

So the fix is to find out which kind the start tile is, mark it properly and remember not to override it later.